requirements.txt
flask==3.0.0
tensorflow>=2.10.0
pillow==9.5.0
numpy==1.24.3
python-dotenv==1.0.0
werkzeug==3.0.1
opencv-python==4.8.1.78
